+++ crates/rome_service/src/configuration/parse/json/rules.rs
@@ -1786,6 +1786,7 @@ impl VisitNode<JsonLanguage> for Nursery {
                 "useAriaPropTypes",
                 "useArrowFunction",
                 "useExhaustiveDependencies",
+                "useGetterReturn",
                 "useGroupedTypeImport",
                 "useHookAtTopLevel",
                 "useImportRestrictions",
@@ -2309,6 +2310,29 @@ impl VisitNode<JsonLanguage> for Nursery {
                         diagnostics,
                     )?;
                     self.use_exhaustive_dependencies = Some(rule_configuration);
+                }
+                _ => {
+                    diagnostics.push(DeserializationDiagnostic::new_incorrect_type(
+                        "object or string",
+                        value.range(),
+                    ));
+                }
+            },
+            "useGetterReturn" => match value {
+                AnyJsonValue::JsonStringValue(_) => {
+                    let mut configuration = RuleConfiguration::default();
+                    self.map_to_known_string(&value, name_text, &mut configuration, diagnostics)?;
+                    self.use_getter_return = Some(configuration);
+                }
+                AnyJsonValue::JsonObjectValue(_) => {
+                    let mut rule_configuration = RuleConfiguration::default();
+                    rule_configuration.map_rule_configuration(
+                        &value,
+                        name_text,
+                        "useGetterReturn",
+                        diagnostics,
+                    )?;
+                    self.use_getter_return = Some(rule_configuration);
                 }
                 _ => {
                     diagnostics.push(DeserializationDiagnostic::new_incorrect_type(